feat(flux-lsp): track changes as internal state
