+++ src/handlers/document_open.rs
@@ -1,8 +1,9 @@
 use crate::cache;
-use crate::handlers::{create_file_diagnostics, RequestHandler};
+use crate::handlers::{create_diagnostics, RequestHandler};
 use crate::protocol::requests::{
     PolymorphicRequest, Request, TextDocumentParams,
 };
+use crate::utils::create_file_node_from_text;
 
 #[derive(Default)]
 pub struct DocumentOpenHandler {}
@@ -20,8 +21,11 @@ impl RequestHandler for DocumentOpenHandler {
             let text = params.text_document.text;
 
             cache::set(uri.clone(), version, text)?;
+            let cv = cache::get(uri.clone())?;
+            let node =
+                create_file_node_from_text(uri.clone(), cv.contents);
 
-            let msg = create_file_diagnostics(uri.clone())?;
+            let msg = create_diagnostics(uri.clone(), node)?;
             let json = msg.to_json()?;
 
             return Ok(Some(json));
