+++ tests/handlers.rs
@@ -6,11 +6,12 @@ use flux_lsp_lib::protocol::notifications::*;
 use flux_lsp_lib::protocol::properties::*;
 use flux_lsp_lib::protocol::requests::*;
 use flux_lsp_lib::protocol::responses::*;
-use flux_lsp_lib::utils;
 
 use std::cell::RefCell;
 use std::collections::HashMap;
+use std::fs;
 use std::rc::Rc;
+use url::Url;
 
 fn flux_fixture_uri(filename: &'static str) -> String {
     let mut pwd = std::env::current_dir().unwrap();
@@ -22,6 +23,29 @@ fn flux_fixture_uri(filename: &'static str) -> String {
     let p = pwd.as_path().to_str().unwrap().to_string();
 
     format!("file://{}", p)
+}
+
+pub fn get_file_contents_from_uri(
+    uri: String,
+) -> Result<String, String> {
+    let url = match Url::parse(uri.as_str()) {
+        Ok(s) => s,
+        Err(e) => {
+            return Err(format!("Failed to get file path: {}", e))
+        }
+    };
+
+    let file_path = match Url::to_file_path(&url) {
+        Ok(s) => s,
+        Err(_) => return Err("Faild to get file_path".to_string()),
+    };
+
+    let contents = match fs::read_to_string(file_path) {
+        Ok(c) => c,
+        Err(e) => return Err(format!("Failed to read file: {}", e)),
+    };
+
+    Ok(contents)
 }
 
 struct TestLogger {}
@@ -40,8 +64,7 @@ fn create_handler() -> Handler {
 }
 
 fn open_file(uri: String, handler: &mut Handler) {
-    let text =
-        utils::get_file_contents_from_uri(uri.clone()).unwrap();
+    let text = get_file_contents_from_uri(uri.clone()).unwrap();
     let did_open_request = Request {
         id: 1,
         method: "textDocument/didOpen".to_string(),
@@ -69,8 +92,7 @@ fn open_file(uri: String, handler: &mut Handler) {
 }
 
 fn close_file(uri: String, handler: &mut Handler) {
-    let text =
-        utils::get_file_contents_from_uri(uri.clone()).unwrap();
+    let text = get_file_contents_from_uri(uri.clone()).unwrap();
     let did_close_request = Request {
         id: 1,
         method: "textDocument/didClose".to_string(),
@@ -278,8 +300,7 @@ fn test_document_open_error() {
 #[test]
 fn test_document_change_ok() {
     let uri = flux_fixture_uri("ok");
-    let text =
-        utils::get_file_contents_from_uri(uri.clone()).unwrap();
+    let text = get_file_contents_from_uri(uri.clone()).unwrap();
 
     let mut handler = create_handler();
 
@@ -327,8 +348,7 @@ fn test_document_change_ok() {
 #[test]
 fn test_document_change_error() {
     let uri = flux_fixture_uri("error");
-    let text =
-        utils::get_file_contents_from_uri(uri.clone()).unwrap();
+    let text = get_file_contents_from_uri(uri.clone()).unwrap();
     let mut handler = create_handler();
 
     with_file_open(uri.clone(), &mut handler, move |handler| {
